Create local client in login
This allows it to be used for other methods and removes some code
duplication.

I had to remove the LocalToken from config in the
`GCP standard login, success` test for it to pass, since otherwise it
would have tried to create a local login.

I added a local login test for Put.