Update vault/api mod to 1.0.3
Also point directly to github.com/hashicorp/vault/api instead of the
broader github.com/hashicorp/vault because the api is versioned on its
own and not using it directly causes dependency issues with trying to
upgrade.

I also ran `go mod tidy` and it removed a lot of mods. I'm assuming
those are for examples?